added login button to home page & general cleanup
