+++ app.js
@@ -8,7 +8,7 @@ var cookieParser = require('cookie-parser');
 
 
 // CUSTOM MODULES
-var db = require('./custom_modules/init_db');
+var db = require('./custom_modules/init-db');
 
 // create express app
 var app = express();
@@ -24,6 +24,7 @@ app.use(cookieParser());
 app.use(logger('dev'));
 app.use(express.json());
 
+
 // ROUTERS
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
@@ -31,6 +32,16 @@ var adminRouter = require('./routes/admin');
 var eventFormRouter = require('./routes/event-form');
 var addEventRouter = require('./routes/add-event');
 var authRouter = require('./routes/auth');
+var authCallbackRouter = require('./routes/auth-callback');
+
+// ROUTES
+app.use('/', indexRouter);
+app.use('/users', usersRouter); // NOT USED!
+app.use('/admin', adminRouter);
+app.use('/event-form', eventFormRouter);
+app.use('/add-event', addEventRouter);
+app.use('/auth', authRouter);
+app.use('/auth-callback', authCallbackRouter);
 
 
 const {
@@ -40,14 +51,6 @@ const {
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'pug');
-
-// ROUTES
-app.use('/', indexRouter);
-app.use('/users', usersRouter); // NOT USED!
-app.use('/admin', adminRouter);
-app.use('/event-form', eventFormRouter);
-app.use('/add-event', addEventRouter);
-app.use('/auth', authRouter); 
 
 // STATIC FILES
 app.use(express.static(path.join(__dirname, 'public')));
