+++ config.js
@@ -36,7 +36,7 @@ module.exports = {
         auth_provider_x509_cert_url: 'https://www.googleapis.com/oauth2/v1/certs',
         client_secret: process.env.GOOGLE_CLIENT_SECRET,
         redirect_uris: [
-            `${baseURL}/auth_callback`
+            `${baseURL}/auth-callback`
         ],
         scopes: [
             'https://www.googleapis.com/auth/userinfo.email',
