+++ public/css/style.css
@@ -1,44 +1,3 @@
-.sidebar-fixed {
-    height: 100vh;
-    width: 270px;
-    -webkit-box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.16), 0 2px 10px 0 rgba(0, 0, 0, 0.12);
-    box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.16), 0 2px 10px 0 rgba(0, 0, 0, 0.12);
-    z-index: 1050;
-    background-color: #fff;
-    padding: 1.5rem;
-    padding-top: 0;
-}
-
-.sidebar-fixed .list-group .active {
-    -webkit-box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.16), 0 2px 10px 0 rgba(0, 0, 0, 0.12);
-    box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.16), 0 2px 10px 0 rgba(0, 0, 0, 0.12);
-    -webkit-border-radius: 5px;
-    border-radius: 5px;
-}
-
-.sidebar-fixed .logo-wrapper {
-    padding: 2.5rem;
-}
-
-.sidebar-fixed .logo-wrapper img {
-    max-height: 50px;
-}
-
-@media (min-width: 1200px) {
-
-    .navbar,
-    .page-footer,
-    main {
-        padding-left: 270px;
-    }
-}
-
-@media (max-width: 1199.98px) {
-    .sidebar-fixed {
-        display: none;
-    }
-}
-
 .btn-rounded {
     border-radius: 25px;
 }
