+++ views/admin.pug
@@ -14,6 +14,7 @@ head
   link(href='css/addons/datatables2.min.css' rel='stylesheet')
   // Your custom styles (optional)
   link(href='css/style.css' rel='stylesheet')
+  link(href='css/style-admin.css' rel='stylesheet')
 
 
   // Style for Google Maps card
